fix(convex): rename identity-signing to identitySigning (Convex disallows hyphens in module paths)
